UPDATE: move api request to group service
