+++ src/main/java/mops/businesslogic/PermissionService.java
@@ -3,7 +3,6 @@ package mops.businesslogic;
 import mops.exception.MopsException;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Set;
 
 /**
@@ -24,10 +23,4 @@ public interface PermissionService {
      * @return gets all roles of that group
      */
     Set<String> fetchRolesInGroup(long groupId) throws MopsException;
-
-    /**
-     * @param account the user credentials
-     * @return a list of groups for that user
-     */
-    List<Group> fetchGroupsForUser(Account account) throws MopsException;
 }
