+++ src/main/java/mops/businesslogic/PermissionServiceDevImpl.java
@@ -4,7 +4,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Set;
 
 @Slf4j
@@ -41,13 +40,5 @@ public class PermissionServiceDevImpl implements PermissionService {
     @Override
     public Set<String> fetchRolesInGroup(long groupId) {
         return Set.of("admin", "editor", "viewer");
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public List<Group> fetchGroupsForUser(Account account) {
-        return null;
     }
 }
